+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
-### Unreleased
+### [v0.59](https://github.com/react4j/react4j/tree/v0.59) (2018-04-04)
+[Full Changelog](https://github.com/react4j/react4j/compare/v0.58...v0.59)
 
 #### Fixed
 * **\[processor\]** Ensure that a react component with a parameterized type will add the parameterized type